Re-enable startup profile. Rework variant filtering logic.
