+++ android/benchmark/src/main/java/io/github/reactivecircus/kstreamlined/android/benchmark/utils.kt
@@ -1,4 +1,4 @@
-package io.github.reactivecircus.benchmark
+package io.github.reactivecircus.kstreamlined.android.benchmark
 
 import androidx.test.platform.app.InstrumentationRegistry
