+++ build-logic/src/main/kotlin/BuildEnvironment.kt
@@ -1,5 +1,3 @@
-package io.github.reactivecircus.kstreamlined.buildlogic
-
 import org.gradle.api.Project
 import org.gradle.api.provider.Provider
 
@@ -8,6 +6,16 @@ val Project.isCiBuild: Boolean
 
 val Project.isIdeBuild: Boolean
     get() = providers.systemProperty("idea.active").orNull == "true"
+
+val Project.isGeneratingBaselineProfile: Boolean
+    get() = gradle.startParameter.taskNames.any {
+        it.contains("generateBaselineProfile", ignoreCase = true)
+    }
+
+val Project.isRunningBenchmark: Boolean
+    get() = gradle.startParameter.taskNames.any {
+        it.contains("connectedBenchmarkReleaseAndroidTest", ignoreCase = true)
+    }
 
 val Project.isAppleSilicon: Boolean
     get() = providers.systemProperty("os.arch").orNull == "aarch64"
