+++ build-logic/src/main/kotlin/ProductFlavorConfigurationAccessors.kt
@@ -1,5 +1,3 @@
-package io.github.reactivecircus.kstreamlined.buildlogic
-
 import org.gradle.api.artifacts.Dependency
 import org.gradle.api.artifacts.dsl.DependencyHandler
