+++ build-logic/src/main/kotlin/io/github/reactivecircus/kstreamlined/buildlogic/KMPBuildLogic.kt
@@ -1,5 +1,6 @@
 package io.github.reactivecircus.kstreamlined.buildlogic
 
+import isAppleSilicon
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
