--- build-logic/src/main/kotlin/io/github/reactivecircus/kstreamlined/buildlogic/VariantExt.kt
@@ -1,15 +1,0 @@
-package io.github.reactivecircus.kstreamlined.buildlogic
-
-import com.android.build.api.variant.BuildConfigField
-import com.android.build.api.variant.ResValue
-import com.android.build.api.variant.Variant
-import java.io.Serializable
-
-fun Variant.addResValue(key: String, type: String, value: String) {
-    resValues.put(makeResValueKey(type, key), ResValue(value))
-}
-
-fun <T : Serializable> Variant.addBuildConfigField(key: String, value: T) {
-    val buildConfigField = BuildConfigField(type = value::class.java.simpleName, value = value, comment = null)
-    buildConfigFields.put(key, buildConfigField)
-}
